[ASV-2219] Converted ThemeAttributeProvider to kotlin class and moved to another package
